Fix untriggered bug of out of bounds access in utf_to_utf functions
The bug could happen after resizing the output string.
Was not triggered due to short string optimizations. Could be triggered
if:
- std::string has no SSO
- input string is utf-32 (wchar), output is utf-8 (char)
- input is very short but contains high unicode chars that
  need 3 or 4 bytes in utf-8.

E.g. one CP as input in UTF-32 that would be 4 CU in UTF-8 would trigger.
After the first resize, the out buffer would contain 2 code units which
was not enough.