Various cleanup.
1. Fix use of inline namespace.

Put inline namespace in all public headers.
Even types that are not meant to be touched by client code are in such
namespace. This is because our main class Dictionary follows the rule
of zero, and has destructor, copy/move ctor and operator= that are
implicitly declared and defined. Those get inlined in client code, and
client code ends up calling copy/move/destruct of all the data members of
Dictionary, actually of the data members of Aff_Data.

2. Simplify CMake by not loading module CTest. This is used only if you
   upload to CDash which we don't do now.

3. Drop unused variable in Dictionary::spell().

4. Fix assignment operators that were returning by value. They should
   return the reference of lhs-object.

5. Fix reference parameters of some functions and constructors.

6. Make default some default constructors, makes them noexcept, too.

7. Improve String_Set in few ways:
   - Add some const and noexcept.
   - Functions that take single char (key_type) now take by value.

8. Drop dependency on tuple.

Tuple is not as efficient as possible when passing by value. That is
because it is not "trivial for ABI purposes" in some implementations,
i.e. not trivially copy-assignable, move-assignable and destructible,
even when the underlying types are. On the other hand, std::pair is ok.

9. No need for string_view in Word_List type.

After we moved all utf-32, the string view is unnecessary,
because member functions are taking key_type by const reference.
String view is more efficient when passed by value.