Improve error-reporting, no more direct usage of cerr.
The library does not use directly any of the standard stream object like
cerr anymore. Instead, in Aff_Parser, any ostream object can be injected
that receives the error message. In the class Dictionary the functions
for loading the dictionary throw as usual but now have much richer error
message. So, now the library does not pollute cout or cerr and it is up
to the user to print the error message via the exception.

Additional function for loading has been added in class Dictionary (only
for internal use) that receives ostream object for error messages.

In the executable CLI tool std::clog is used everywhere instead of
std::cerr and explicit flushes with std::endl are used where needed.